pyproject.toml
[project]
name = "ai-agents-lab"
version = "0.1.0"
description = "AI Agents Learning Lab - Production patterns for LLM agents"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.15.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["shared*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
